+++ drift/lib/internal/versioned_schema.dart
@@ -98,6 +98,13 @@ abstract base class VersionedSchema {
     required MigrationStepWithVersion steps,
   }) async {
     final database = migrator.database;
+    if (from > to) {
+      throw StateError(
+        "runMigrationSteps was asked to downgrade from versions $from to $to. "
+        'Migration steps can only upgrade. This might happen if users '
+        'downgrade app versions.',
+      );
+    }
 
     for (var target = from; target < to;) {
       final newVersion = await steps(target, database);